requirements.txt
flask~=3.1.1
flask_sqlalchemy
SQLAlchemy~=2.0.41
Werkzeug~=3.1.3
flask-cors~=6.0.1
PyJWT~=2.10.1
python-dotenv~=1.1.1
Flask-Admin~=1.6.1